resolved most code duplications
