+++ tests/test_reshuffle.py
@@ -27,7 +27,7 @@ import numpy.testing as nptest
 import tempfile
 from smos.smos_ic.reshuffle_ic import main
 import glob
-from smos.smos_ic.interface import SMOSTs
+from smos.smos_ic.interface_ic import SMOSTs
 from datetime import timedelta
